code changes for extension resources
