+++ src/AutoRest.CSharp/Mgmt/AutoRest/MgmtOutputLibrary.cs
@@ -237,10 +237,27 @@ namespace AutoRest.CSharp.Mgmt.AutoRest
             {
                 var resourceType = operationGroup.IsTupleResource(_context) ? ResourceType.Tuple : ResourceType.Default;
                 var childOperationGroups = _childNonResourceOperationGroups.GetValueOrDefault(operationGroup.ResourceType(_mgmtConfiguration));
-                _resourceOperations[resourceType].Add(operationGroup, new ResourceOperation(operationGroup, _context, childOperationGroups));
+                var resourceOperation = new ResourceOperation(operationGroup, _context, childOperationGroups);
+                // validate to ensure that all the resource operations here have unique names
+                ValidateUniqueName(_resourceOperations, resourceOperation);
+                _resourceOperations[resourceType].Add(operationGroup, resourceOperation);
             }
 
             return _resourceOperations;
+        }
+
+        private static void ValidateUniqueName<T, V>(IDictionary<ResourceType, V> map, T value) where T : ResourceOperation where V : IDictionary<OperationGroup, T>
+        {
+            // we need to iterate over the existing items (including Default resource type and Tuple resource type)
+            // to see if there are already any resource operations are returning the same resource as this new one
+            foreach (var pair in map.SelectMany(p => p.Value))
+            {
+                var existing = pair.Value;
+                if (value.Type.Name.Equals(existing.Type.Name))
+                {
+                    throw new Exception($"Operation Group {existing.OperationGroup.Key} and {value.OperationGroup.Key} return the same resource {existing.ResourceName}, please consider mark these operation groups as extension resource in the `operation-group-is-extension` section");
+                }
+            }
         }
 
         private Dictionary<ResourceType, Dictionary<OperationGroup, ResourceContainer>> EnsureResourceContainers()
@@ -258,7 +275,10 @@ namespace AutoRest.CSharp.Mgmt.AutoRest
                 if (!operationGroup.IsSingletonResource(_context.Configuration.MgmtConfiguration))
                 {
                     var resourceType = operationGroup.IsTupleResource(_context) ? ResourceType.Tuple : ResourceType.Default;
-                    _resourceContainers[resourceType].Add(operationGroup, new ResourceContainer(operationGroup, _context));
+                    var resourceContainer = new ResourceContainer(operationGroup, _context);
+                    // validate to ensure that all the resource container here have unique names
+                    ValidateUniqueName(_resourceContainers, resourceContainer);
+                    _resourceContainers[resourceType].Add(operationGroup, resourceContainer);
                 }
             }
