+++ src/AutoRest.CSharp/Mgmt/Output/Resource.cs
@@ -1,6 +1,8 @@
 ﻿// Copyright (c) Microsoft Corporation. All rights reserved.
 // Licensed under the MIT License.
 
+using System;
+using System.Linq;
 using AutoRest.CSharp.Input;
 using AutoRest.CSharp.Mgmt.Decorator;
 using AutoRest.CSharp.Output.Builders;
@@ -14,7 +16,15 @@ namespace AutoRest.CSharp.Mgmt.Output
             : base(context)
         {
             OperationGroup = operationGroup;
-            DefaultName = operationGroup.Resource(context.Configuration.MgmtConfiguration);
+            // check if this is an extension resource, if so, we need to append the name of its parent to this resource name
+            var isExtension = operationGroup.IsExtensionResource(context.Configuration.MgmtConfiguration);
+            string parentValue = "";
+            if (isExtension)
+            {
+                var parentArr = operationGroup.ParentResourceType(context.Configuration.MgmtConfiguration).Split('/');
+                parentValue = FirstCharToUpper(parentArr.Last());
+            }
+            DefaultName = operationGroup.Resource(context.Configuration.MgmtConfiguration) + parentValue;
             Description = BuilderHelpers.EscapeXmlDescription(
                 $"A Class representing a {DefaultName} along with the instance operations that can be performed on it.");
         }
@@ -26,5 +36,13 @@ namespace AutoRest.CSharp.Mgmt.Output
         protected override string DefaultAccessibility => "public";
 
         public string Description { get; }
+
+        private static string FirstCharToUpper(string input) =>
+        input switch
+        {
+            null => throw new ArgumentNullException(nameof(input)),
+            "" => throw new ArgumentException($"{nameof(input)} cannot be empty", nameof(input)),
+            _ => input.First().ToString().ToUpper() + input.Substring(1)
+        };
     }
 }
