+++ test/TestProjects/MgmtMultipleParentResource/readme.md
@@ -7,15 +7,10 @@ azure-arm: true
 require: $(this-folder)/../../../readme.md
 input-file: $(this-folder)/mgmtMultipleParentResource.json
 namespace: MgmtMultipleParentResource
-# payload-flattening-threshold: 2
-# operation-group-to-resource-type:
-#    VirtualMachineExtensionImages: Microsoft.Compute/locations/publishers/vmextension
-# operation-group-to-resource:
-#    VirtualMachineExtensionImages: VirtualMachineExtensionImage
 operation-group-to-parent:
    Children: Microsoft.Compute/parents/subParents
    AnotherChildren: Microsoft.Compute/anotherParents
-# operation-group-is-tuple: VirtualMachineExtensionImages
+operation-group-is-extension: Children;AnotherChildren
 modelerfour:
   lenient-model-deduplication: true
 ```
